+++ R/gg_contour_filled.R
@@ -1,34 +1,20 @@
-#' @title Blank ggplot
+#' @title contour_filled2 ggplot
 #'
-#' @description Create a blank ggplot with a wrapper around the ggplot2::geom_blank function.
+#' @description Create a bin2d ggplot with a wrapper around the ggplot2::geom_contour_filled function.
 #' @param data A data frame or tibble.
 #' @param x Unquoted x aesthetic variable.
 #' @param y Unquoted y aesthetic variable.
-#' @param col Unquoted col and fill aesthetic variable.
+#' @param z Unquoted z aesthetic variable
 #' @param facet Unquoted facet aesthetic variable.
 #' @param facet2 Unquoted second facet variable.
 #' @param group Unquoted group aesthetic variable.
-#' @param xmin Unquoted xmin aesthetic variable.
-#' @param xmax Unquoted xmax aesthetic variable.
-#' @param xend Unquoted xend aesthetic variable.
-#' @param xlower Unquoted xlower aesthetic variable.
-#' @param xupper Unquoted xupper aesthetic variable.
-#' @param xmiddle Unquoted xmiddle aesthetic variable.
-#' @param ymin Unquoted ymin aesthetic variable.
-#' @param ymax Unquoted ymax aesthetic variable.
-#' @param yend Unquoted yend aesthetic variable.
-#' @param ylower Unquoted ylower aesthetic variable.
-#' @param yupper Unquoted yupper aesthetic variable.
-#' @param ymiddle Unquoted ymiddle aesthetic variable.
-#' @param sample Unquoted sample aesthetic variable.
-#' @param label Unquoted label aesthetic variable.
-#' @param subgroup Unquoted subgroup aesthetic variable.
 #' @param stat Statistical transformation. A character string (e.g. "identity").
 #' @param position Position adjustment. Either a character string (e.g."identity"), or a function (e.g. ggplot2::position_identity()).
 #' @param coord A coordinate function from ggplot2 (e.g. ggplot2::coord_cartesian()).
 #' @param pal Colours to use. A character vector of hex codes (or names).
 #' @param pal_na Colour to use for NA values. A character vector of a hex code (or name).
-#' @param ... Other arguments passed to the ggplot2::geom_blank function.
+#' @param alpha Opacity. A number between 0 and 1.
+#' @param ... Other arguments passed to the ggplot2::geom_contour_filled function.
 #' @param title Title string.
 #' @param subtitle Subtitle string.
 #' @param x_breaks A scales::breaks_* function (e.g. scales::breaks_pretty()), or a vector of breaks.
@@ -78,47 +64,29 @@
 #' @return A ggplot object.
 #' @export
 #' @examples
-#' library(dplyr)
-#' library(palmerpenguins)
+#' ggplot2::faithfuld |>
+#'  gg_contour_filled(
+#'    x = waiting,
+#'    y = eruptions,
+#'    z = density,
+#'    bins = 8,
+#'    y_limits = c(NA, NA)
+#'  )
 #'
-#' penguins |>
-#'   gg_blank(
-#'     x = flipper_length_mm,
-#'     y = body_mass_g,
-#'     col = sex,
-#'     facet = species,
-#'     col_labels = stringr::str_to_sentence,
-#'     pal = c("#1B9E77", "#9E361B")
-#'   )
-#'
-gg_blank <- function(
+gg_contour_filled <- function(
     data = NULL,
     x = NULL,
     y = NULL,
-    col = NULL,
+    z = NULL,
     facet = NULL,
     facet2 = NULL,
     group = NULL,
-    xmin = NULL,
-    xmax = NULL,
-    xend = NULL,
-    xlower = NULL,
-    xupper = NULL,
-    xmiddle = NULL,
-    ymin = NULL,
-    ymax = NULL,
-    yend = NULL,
-    ylower = NULL,
-    yupper = NULL,
-    ymiddle = NULL,
-    sample = NULL,
-    label = NULL,
-    subgroup = NULL,
-    stat = "identity",
+    stat = "contour_filled",
     position = "identity",
     coord = ggplot2::coord_cartesian(clip = "off"),
     pal = NULL,
     pal_na = pal_grey,
+    alpha = 1,
     ...,
     title = NULL,
     subtitle = NULL,
@@ -173,159 +141,43 @@ gg_blank <- function(
   #quote
   x <- rlang::enquo(x)
   y <- rlang::enquo(y)
-  col <- rlang::enquo(col)
+  z <- rlang::enquo(z)
   facet <- rlang::enquo(facet)
   facet2 <- rlang::enquo(facet2)
   group <- rlang::enquo(group)
-
-  xmin <- rlang::enquo(xmin)
-  xmax <- rlang::enquo(xmax)
-  xend <- rlang::enquo(xend)
-  xlower <- rlang::enquo(xlower)
-  xupper <- rlang::enquo(xupper)
-  xmiddle <- rlang::enquo(xmiddle)
-
-  ymin <- rlang::enquo(ymin)
-  ymax <- rlang::enquo(ymax)
-  yend <- rlang::enquo(yend)
-  ylower <- rlang::enquo(ylower)
-  yupper <- rlang::enquo(yupper)
-  ymiddle <- rlang::enquo(ymiddle)
-
-  sample <- rlang::enquo(sample)
-  label <- rlang::enquo(label)
-  subgroup <- rlang::enquo(subgroup)
 
   #ungroup
   data <- data %>%
     dplyr::ungroup() %>%
     dplyr::mutate(dplyr::across(
-      c(!!x, !!y,
-        !!xmin, !!ymin,
-        !!xmax, !!ymax,
-        !!xend, !!yend,
-        !!xlower, !!xmiddle, !!xupper, !!ylower, !!ymiddle, !!yupper,
-        !!col
-      ),
+      c(!!x, !!y),
       na_if_inf))
 
   #get classes
-  if (stat == "sf") {
-    x_null <- TRUE
-    x_factor <- FALSE
-    x_forcat <- FALSE
-    x_numeric <- FALSE
-    x_date <- FALSE
-    x_datetime <- FALSE
-    x_time <- FALSE
+  x_null <- rlang::quo_is_null(x)
+  x_forcat <- is.character(rlang::eval_tidy(x, data)) | is.factor(rlang::eval_tidy(x, data)) | is.logical(rlang::eval_tidy(x, data))
+  x_numeric <- is.numeric(rlang::eval_tidy(x, data))
+  x_date <- lubridate::is.Date(rlang::eval_tidy(x, data))
+  x_datetime <- lubridate::is.POSIXct(rlang::eval_tidy(x, data))
+  x_time <- hms::is_hms(rlang::eval_tidy(x, data))
 
-    y_null <- TRUE
-    y_factor <- FALSE
-    y_forcat <- FALSE
-    y_numeric <- FALSE
-    y_date <- FALSE
-    y_datetime <- FALSE
-    y_time <- FALSE
-  }
-  else {
-    x_null <- rlang::quo_is_null(x) & rlang::quo_is_null(xmin) & rlang::quo_is_null(xmax) & rlang::quo_is_null(xend)
-    x_forcat <- is.character(rlang::eval_tidy(x, data)) | is.factor(rlang::eval_tidy(x, data)) | is.logical(rlang::eval_tidy(x, data))
-    x_numeric <- {
-      is.numeric(rlang::eval_tidy(x, data)) |
-        is.numeric(rlang::eval_tidy(xmin, data)) |
-        is.numeric(rlang::eval_tidy(xmax, data)) |
-        is.numeric(rlang::eval_tidy(xend, data))
-    }
-    x_date <- {
-      lubridate::is.Date(rlang::eval_tidy(x, data)) |
-        lubridate::is.Date(rlang::eval_tidy(xmin, data)) |
-        lubridate::is.Date(rlang::eval_tidy(xmax, data)) |
-        lubridate::is.Date(rlang::eval_tidy(xend, data))
-    }
-    x_datetime <- {
-      lubridate::is.POSIXct(rlang::eval_tidy(x, data)) |
-        lubridate::is.POSIXct(rlang::eval_tidy(xmin, data)) |
-        lubridate::is.POSIXct(rlang::eval_tidy(xmax, data)) |
-        lubridate::is.POSIXct(rlang::eval_tidy(xend, data))
-    }
-    x_time <- {
-      hms::is_hms(rlang::eval_tidy(x, data)) |
-        hms::is_hms(rlang::eval_tidy(xmin, data)) |
-        hms::is_hms(rlang::eval_tidy(xmax, data)) |
-        hms::is_hms(rlang::eval_tidy(xend, data))
-    }
+  y_null <- rlang::quo_is_null(y)
+  y_forcat <- is.character(rlang::eval_tidy(y, data)) | is.factor(rlang::eval_tidy(y, data)) | is.logical(rlang::eval_tidy(y, data))
+  y_numeric <- is.numeric(rlang::eval_tidy(y, data))
+  y_date <- lubridate::is.Date(rlang::eval_tidy(y, data))
+  y_datetime <- lubridate::is.POSIXct(rlang::eval_tidy(y, data))
+  y_time <- hms::is_hms(rlang::eval_tidy(y, data))
 
-    y_null <- rlang::quo_is_null(y) & rlang::quo_is_null(ymin) & rlang::quo_is_null(ymax) & rlang::quo_is_null(yend)
-    y_forcat <- is.character(rlang::eval_tidy(y, data)) | is.factor(rlang::eval_tidy(y, data)) | is.logical(rlang::eval_tidy(y, data))
-    y_numeric <- {
-      is.numeric(rlang::eval_tidy(y, data)) |
-        is.numeric(rlang::eval_tidy(ymin, data)) |
-        is.numeric(rlang::eval_tidy(ymax, data)) |
-        is.numeric(rlang::eval_tidy(yend, data))
-    }
-    y_date <- {
-      lubridate::is.Date(rlang::eval_tidy(y, data)) |
-        lubridate::is.Date(rlang::eval_tidy(ymin, data)) |
-        lubridate::is.Date(rlang::eval_tidy(ymax, data)) |
-        lubridate::is.Date(rlang::eval_tidy(yend, data))
-    }
-    y_datetime <- {
-      lubridate::is.POSIXct(rlang::eval_tidy(y, data)) |
-        lubridate::is.POSIXct(rlang::eval_tidy(ymin, data)) |
-        lubridate::is.POSIXct(rlang::eval_tidy(ymax, data)) |
-        lubridate::is.POSIXct(rlang::eval_tidy(yend, data))
-    }
-    y_time <- {
-      hms::is_hms(rlang::eval_tidy(y, data)) |
-        hms::is_hms(rlang::eval_tidy(ymin, data)) |
-        hms::is_hms(rlang::eval_tidy(ymax, data)) |
-        hms::is_hms(rlang::eval_tidy(yend, data))
-    }
-  }
-
-  if (stat %in% c("bin2d", "bin_2d", "binhex")) {
-    col_null <- TRUE
-    col_factor <- FALSE
-    col_forcat <- FALSE
-    col_numeric <- FALSE
-    col_date <- FALSE
-    col_datetime <- FALSE
-    col_time <- FALSE
-  }
-  else {
-    col_null <- rlang::quo_is_null(col)
-    col_factor <- is.factor(rlang::eval_tidy(col, data))
-    col_forcat <- is.character(rlang::eval_tidy(col, data)) | is.factor(rlang::eval_tidy(col, data)) | is.logical(rlang::eval_tidy(col, data))
-    col_numeric <- is.numeric(rlang::eval_tidy(col, data))
-    col_date <- lubridate::is.Date(rlang::eval_tidy(col, data))
-    col_datetime <- lubridate::is.POSIXct(rlang::eval_tidy(col, data))
-    col_time <- hms::is_hms(rlang::eval_tidy(col, data))
-  }
+  col_null <- TRUE
+  col_factor <- FALSE
+  col_forcat <- FALSE
+  col_numeric <- FALSE
+  col_date <- FALSE
+  col_datetime <- FALSE
+  col_time <- FALSE
 
   facet_null <- rlang::quo_is_null(facet)
   facet2_null <- rlang::quo_is_null(facet2)
-
-  # if (rlang::is_null(alpha)) {
-  #   # geometry_type <- unique(sf::st_geometry_type(data))
-  #   # if (length(geometry_type) > 1) geometry_type <- "GEOMETRY"
-  #   geometry_type <-
-  #     stringr::str_remove(attributes(sf::st_geometry(data))$class[1], "sfc_")
-  #
-  #   if (geometry_type %in% c("POINT",
-  #                            "MULTIPOINT",
-  #                            "LINESTRING",
-  #                            "MULTILINESTRING",
-  #                            "CIRCULARSTRING",
-  #                            "COMPOUNDCURVE",
-  #                            "MULTICURVE",
-  #                            "CURVE")) {
-  #     alpha <- 1
-  #   }
-  #   else {
-  #     if (col_null) alpha <- 0
-  #     else alpha <- 0.9
-  #   }
-  # }
 
   ##############################################################################
   #Generic code: part 1 (adjust for gg_sf)
@@ -372,113 +224,37 @@ gg_blank <- function(
   ##############################################################################
 
   ###make plot
-  if (stat == "sf") {
-    if (!col_null) {
-      plot <- data %>%
-        ggplot2::ggplot(mapping = ggplot2::aes(
-          geometry = sf::st_geometry(data),
-          col = !!col,
-          fill = !!col,
-          group = !!group
-        ))
-    }
-    else if (col_null) {
-      plot <- data %>%
-        ggplot2::ggplot(mapping = ggplot2::aes(
-          geometry = sf::st_geometry(data),
-          # col = "",
-          # fill = "",
-          group = !!group
-        ))
-    }
-  }
-  else if (stat %in% c("bin2d", "bin_2d", "binhex")) {
+  if (stat %in% c("bin2d", "bin_2d", "binhex", "contour_filled", "density_2d_filled")) {
     if (!x_null & !y_null) {
       plot <- data %>%
         ggplot2::ggplot(mapping = ggplot2::aes(
           x = !!x,
           y = !!y,
-          group = !!group,
-          xmin = !!xmin,
-          xmax = !!xmax,
-          xend = !!xend,
-          xlower = !!xlower,
-          xupper = !!xupper,
-          xmiddle = !!xmiddle,
-          ymin = !!ymin,
-          ymax = !!ymax,
-          yend = !!yend,
-          lower = !!ylower,
-          upper = !!yupper,
-          middle = !!ymiddle,
-          sample = !!sample,
-          label = !!label,
-          subgroup = !!subgroup
+          z = !!z,
+          group = !!group
         ))
     }
     else if (!x_null & y_null) {
       plot <- data %>%
         ggplot2::ggplot(mapping = ggplot2::aes(
           x = !!x,
-          group = !!group,
-          xmin = !!xmin,
-          xmax = !!xmax,
-          xend = !!xend,
-          xlower = !!xlower,
-          xupper = !!xupper,
-          xmiddle = !!xmiddle,
-          ymin = !!ymin,
-          ymax = !!ymax,
-          yend = !!yend,
-          lower = !!ylower,
-          upper = !!yupper,
-          middle = !!ymiddle,
-          sample = !!sample,
-          label = !!label,
-          subgroup = !!subgroup
+          z = !!z,
+          group = !!group
         ))
     }
     else if (x_null & !y_null) {
       plot <- data %>%
         ggplot2::ggplot(mapping = ggplot2::aes(
           y = !!y,
-          group = !!group,
-          xmin = !!xmin,
-          xmax = !!xmax,
-          xend = !!xend,
-          xlower = !!xlower,
-          xupper = !!xupper,
-          xmiddle = !!xmiddle,
-          ymin = !!ymin,
-          ymax = !!ymax,
-          yend = !!yend,
-          lower = !!ylower,
-          upper = !!yupper,
-          middle = !!ymiddle,
-          sample = !!sample,
-          label = !!label,
-          subgroup = !!subgroup
+          z = !!z,
+          group = !!group
         ))
     }
     else if (x_null & y_null) {
       plot <- data %>%
         ggplot2::ggplot(mapping = ggplot2::aes(
-          group = !!group,
-          xmin = !!xmin,
-          xmax = !!xmax,
-          xend = !!xend,
-          xlower = !!xlower,
-          xupper = !!xupper,
-          xmiddle = !!xmiddle,
-          ymin = !!ymin,
-          ymax = !!ymax,
-          yend = !!yend,
-          lower = !!ylower,
-          upper = !!yupper,
-          middle = !!ymiddle,
-          sample = !!sample,
-          label = !!label,
-          subgroup = !!subgroup
+          z = !!z,
+          group = !!group
         ))
     }
   }
@@ -489,24 +265,10 @@ gg_blank <- function(
           ggplot2::ggplot(mapping = ggplot2::aes(
             x = !!x,
             y = !!y,
+            z = !!z,
             col = !!col,
             fill = !!col,
-            group = !!group,
-            xmin = !!xmin,
-            xmax = !!xmax,
-            xend = !!xend,
-            xlower = !!xlower,
-            xupper = !!xupper,
-            xmiddle = !!xmiddle,
-            ymin = !!ymin,
-            ymax = !!ymax,
-            yend = !!yend,
-            lower = !!ylower,
-            upper = !!yupper,
-            middle = !!ymiddle,
-            sample = !!sample,
-            label = !!label,
-            subgroup = !!subgroup
+            group = !!group
           ))
       }
       else if (col_null) {
@@ -514,24 +276,10 @@ gg_blank <- function(
           ggplot2::ggplot(mapping = ggplot2::aes(
             x = !!x,
             y = !!y,
+            z = !!z,
             # col = "",
             # fill = "",
-            group = !!group,
-            xmin = !!xmin,
-            xmax = !!xmax,
-            xend = !!xend,
-            xlower = !!xlower,
-            xupper = !!xupper,
-            xmiddle = !!xmiddle,
-            ymin = !!ymin,
-            ymax = !!ymax,
-            yend = !!yend,
-            lower = !!ylower,
-            upper = !!yupper,
-            middle = !!ymiddle,
-            sample = !!sample,
-            label = !!label,
-            subgroup = !!subgroup
+            group = !!group
           ))
       }
     }
@@ -540,48 +288,20 @@ gg_blank <- function(
         plot <- data %>%
           ggplot2::ggplot(mapping = ggplot2::aes(
             x = !!x,
+            z = !!z,
             col = !!col,
             fill = !!col,
-            group = !!group,
-            xmin = !!xmin,
-            xmax = !!xmax,
-            xend = !!xend,
-            xlower = !!xlower,
-            xupper = !!xupper,
-            xmiddle = !!xmiddle,
-            ymin = !!ymin,
-            ymax = !!ymax,
-            yend = !!yend,
-            lower = !!ylower,
-            upper = !!yupper,
-            middle = !!ymiddle,
-            sample = !!sample,
-            label = !!label,
-            subgroup = !!subgroup
+            group = !!group
           ))
       }
       else if (col_null) {
         plot <- data %>%
           ggplot2::ggplot(mapping = ggplot2::aes(
             x = !!x,
+            z = !!z,
             # col = "",
             # fill = "",
-            group = !!group,
-            xmin = !!xmin,
-            xmax = !!xmax,
-            xend = !!xend,
-            xlower = !!xlower,
-            xupper = !!xupper,
-            xmiddle = !!xmiddle,
-            ymin = !!ymin,
-            ymax = !!ymax,
-            yend = !!yend,
-            lower = !!ylower,
-            upper = !!yupper,
-            middle = !!ymiddle,
-            sample = !!sample,
-            label = !!label,
-            subgroup = !!subgroup
+            group = !!group
           ))
       }
     }
@@ -590,48 +310,20 @@ gg_blank <- function(
         plot <- data %>%
           ggplot2::ggplot(mapping = ggplot2::aes(
             y = !!y,
+            z = !!z,
             col = !!col,
             fill = !!col,
-            group = !!group,
-            xmin = !!xmin,
-            xmax = !!xmax,
-            xend = !!xend,
-            xlower = !!xlower,
-            xupper = !!xupper,
-            xmiddle = !!xmiddle,
-            ymin = !!ymin,
-            ymax = !!ymax,
-            yend = !!yend,
-            lower = !!ylower,
-            upper = !!yupper,
-            middle = !!ymiddle,
-            sample = !!sample,
-            label = !!label,
-            subgroup = !!subgroup
+            group = !!group
           ))
       }
       else if (col_null) {
         plot <- data %>%
           ggplot2::ggplot(mapping = ggplot2::aes(
             y = !!y,
+            z = !!z,
             # col = "",
             # fill = "",
-            group = !!group,
-            xmin = !!xmin,
-            xmax = !!xmax,
-            xend = !!xend,
-            xlower = !!xlower,
-            xupper = !!xupper,
-            xmiddle = !!xmiddle,
-            ymin = !!ymin,
-            ymax = !!ymax,
-            yend = !!yend,
-            lower = !!ylower,
-            upper = !!yupper,
-            middle = !!ymiddle,
-            sample = !!sample,
-            label = !!label,
-            subgroup = !!subgroup
+            group = !!group
           ))
       }
     }
@@ -639,60 +331,33 @@ gg_blank <- function(
       if (!col_null) {
         plot <- data %>%
           ggplot2::ggplot(mapping = ggplot2::aes(
+            z = !!z,
             col = !!col,
             fill = !!col,
-            group = !!group,
-            xmin = !!xmin,
-            xmax = !!xmax,
-            xend = !!xend,
-            xlower = !!xlower,
-            xupper = !!xupper,
-            xmiddle = !!xmiddle,
-            ymin = !!ymin,
-            ymax = !!ymax,
-            yend = !!yend,
-            lower = !!ylower,
-            upper = !!yupper,
-            middle = !!ymiddle,
-            sample = !!sample,
-            label = !!label,
-            subgroup = !!subgroup
+            group = !!group
           ))
       }
       else if (col_null) {
         plot <- data %>%
           ggplot2::ggplot(mapping = ggplot2::aes(
+            z = !!z,
             # col = "",
             # fill = "",
-            group = !!group,
-            xmin = !!xmin,
-            xmax = !!xmax,
-            xend = !!xend,
-            xlower = !!xlower,
-            xupper = !!xupper,
-            xmiddle = !!xmiddle,
-            ymin = !!ymin,
-            ymax = !!ymax,
-            yend = !!yend,
-            lower = !!ylower,
-            upper = !!yupper,
-            middle = !!ymiddle,
-            sample = !!sample,
-            label = !!label,
-            subgroup = !!subgroup
+            group = !!group
           ))
       }
     }
   }
 
-  if (col_null & !stat %in% c("bin2d", "bin_2d", "binhex")) {
+  if (col_null & !stat %in% c("bin2d", "bin_2d", "binhex", "contour_filled", "density_2d_filled")) {
     if (rlang::is_null(pal)) pal <-  pal_blue
     else pal <- as.vector(pal[1])
 
     plot <- plot +
-      ggplot2::geom_blank(
+      ggplot2::geom_contour_filled(
         stat = stat,
         position = position,
+        alpha = alpha,
         col = pal,
         fill = pal,
         ...
@@ -702,9 +367,10 @@ gg_blank <- function(
   }
   else {
     plot <- plot +
-      ggplot2::geom_blank(
+      ggplot2::geom_contour_filled(
         stat = stat,
         position = position,
+        alpha = alpha,
         ...
       ) +
       coord +
@@ -1324,21 +990,34 @@ gg_blank <- function(
   }
 
   #make col scale
-  if (!col_null | stat %in% c("bin2d", "bin_2d", "binhex")) {
-    if (stat %in% c("bin2d", "bin_2d", "binhex")) {
-      if (!rlang::is_null(plot_build$plot$labels$colour)) {
-        col_vctr <- dplyr::pull(plot_data, rlang::as_name(plot_build$plot$labels$colour[1]))
-      }
-      else if (!rlang::is_null(plot_build$plot$labels$fill)) {
+  if (!col_null | stat %in% c("bin2d", "bin_2d", "binhex", "contour_filled", "density_2d_filled")) {
+    if (stat %in% c("bin2d", "bin_2d", "binhex", "contour_filled", "density_2d_filled")) {
+      if (!rlang::is_null(plot_build$plot$labels$fill)) {
         col_vctr <- dplyr::pull(plot_data, rlang::as_name(plot_build$plot$labels$fill[1]))
       }
+      else if (!rlang::is_null(plot_build$plot$labels$colour)) {
+        col_vctr <- dplyr::pull(plot_data, rlang::as_name(plot_build$plot$labels$colour[1]))
+      }
+
     }
     else if (!col_null) {
       col_vctr <- data %>%
         dplyr::pull(!!col)
     }
 
-    if (col_forcat) {
+    if (rlang::is_null(col_legend_place)) {
+      if (col_numeric | col_date | col_datetime | col_time) col_legend_place <- "right"
+      else if (stat %in% c("bin2d", "binhex", "density_2d_filled", "contour_filled")) col_legend_place <- "right"
+      else col_legend_place <- "bottom"
+    }
+
+    if (col_legend_place == "b") col_legend_place <- "bottom"
+    if (col_legend_place == "t") col_legend_place <- "top"
+    if (col_legend_place == "l") col_legend_place <- "left"
+    if (col_legend_place == "r") col_legend_place <- "right"
+    if (col_legend_place == "n") col_legend_place <- "none"
+
+    if (col_forcat | stat %in% c("contour_filled", "density_2d_filled")) {
       if (!rlang::is_null(col_limits)) col_n <- length(col_limits)
       else if (!rlang::is_null(col_breaks)) col_n <- length(col_breaks)
       else {
@@ -1348,7 +1027,10 @@ gg_blank <- function(
           col_n <- length(col_unique[!is.na(col_unique)])
         }
       }
-      if (rlang::is_null(pal)) pal <- pal_hue[1:col_n]
+      if (rlang::is_null(pal)) {
+        if (stat %in% c("contour_filled", "density_2d_filled")) pal <- viridis::viridis(col_n)
+        else pal <- pal_hue[1:col_n]
+      }
       else if (rlang::is_null(names(pal))) pal <- pal[1:col_n]
 
       if (y_numeric | y_date | y_datetime | y_time) {
@@ -1509,12 +1191,13 @@ gg_blank <- function(
     }
   }
   if (rlang::is_null(col_title)) {
-    if (!rlang::is_null(plot_build$plot$labels$colour)) {
-      col_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$colour[1]), titles)
-    }
-    else if (!rlang::is_null(plot_build$plot$labels$fill)) {
+    if (!rlang::is_null(plot_build$plot$labels$fill)) {
       col_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$fill[1]), titles)
     }
+    else if (!rlang::is_null(plot_build$plot$labels$colour)) {
+      col_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$colour[1]), titles)
+    }
+
   }
 
   plot <- plot +
@@ -1523,7 +1206,7 @@ gg_blank <- function(
       subtitle = subtitle,
       caption = caption)
 
-  if (!col_null | stat %in% c("bin2d", "bin_2d", "binhex")) {
+  if (!col_null | stat %in% c("bin2d", "bin_2d", "binhex", "contour_filled", "density_2d_filled")) {
     plot <- plot +
       ggplot2::labs(
         col = col_title,
@@ -1575,42 +1258,32 @@ gg_blank <- function(
   }
 
   #Adjust legend
-  if (rlang::is_null(col_legend_place)) {
-    if (col_numeric | col_date | col_datetime | col_time) col_legend_place <- "right"
-    else if (stat %in% c("bin2d", "binhex", "density_2d_filled", "contour_filled")) col_legend_place <- "right"
-    else col_legend_place <- "bottom"
-  }
-
-  if (col_legend_place == "b") col_legend_place <- "bottom"
-  if (col_legend_place == "t") col_legend_place <- "top"
-  if (col_legend_place == "l") col_legend_place <- "left"
-  if (col_legend_place == "r") col_legend_place <- "right"
-  if (col_legend_place == "n") col_legend_place <- "none"
-
-  if (col_legend_place %in% c("top", "bottom")) {
-    plot <- plot +
-      ggplot2::theme(legend.position = col_legend_place) +
-      ggplot2::theme(legend.direction = "horizontal") +
-      ggplot2::theme(legend.justification = "left") +
-      ggplot2::theme(legend.box.margin = ggplot2::margin(t = -2.5)) +
-      ggplot2::theme(legend.text = ggplot2::element_text(margin = ggplot2::margin(r = 7.5))) +
-      ggplot2::theme(legend.title = ggplot2::element_text(margin = ggplot2::margin(t = 5)))
-
-    if (col_numeric | stat %in% c("bin2d", "bin_2d", "binhex")) {
+  if (!rlang::is_null(col_legend_place)){
+    if (col_legend_place %in% c("top", "bottom")) {
       plot <- plot +
-        ggplot2::theme(legend.key.width = grid::unit(0.66, "cm")) +
-        ggplot2::theme(legend.text.align = 0.5)
+        ggplot2::theme(legend.position = col_legend_place) +
+        ggplot2::theme(legend.direction = "horizontal") +
+        ggplot2::theme(legend.justification = "left") +
+        ggplot2::theme(legend.box.margin = ggplot2::margin(t = -2.5)) +
+        ggplot2::theme(legend.text = ggplot2::element_text(margin = ggplot2::margin(r = 7.5))) +
+        ggplot2::theme(legend.title = ggplot2::element_text(margin = ggplot2::margin(t = 5)))
+
+      if (col_numeric | stat %in% c("bin2d", "bin_2d", "binhex")) {
+        plot <- plot +
+          ggplot2::theme(legend.key.width = grid::unit(0.66, "cm")) +
+          ggplot2::theme(legend.text.align = 0.5)
+      }
     }
-  }
-  else if (col_legend_place %in% c("left", "right")) {
-    plot <- plot +
-      ggplot2::theme(legend.position = col_legend_place) +
-      ggplot2::theme(legend.direction = "vertical") +
-      ggplot2::theme(legend.justification = "left")
-  }
-  else if (col_legend_place == "none") {
-    plot <- plot +
-      ggplot2::guides(col = "none", fill = "none")
+    else if (col_legend_place %in% c("left", "right")) {
+      plot <- plot +
+        ggplot2::theme(legend.position = col_legend_place) +
+        ggplot2::theme(legend.direction = "vertical") +
+        ggplot2::theme(legend.justification = "left")
+    }
+    else if (col_legend_place == "none") {
+      plot <- plot +
+        ggplot2::guides(col = "none", fill = "none")
+    }
   }
 
   #remove gridlines as per x_gridlines and y_gridlines. Guess if NULL
@@ -1651,3 +1324,4 @@ gg_blank <- function(
   #return beautiful plot
   return(plot)
 }
+